fix(security): resolve remaining CodeQL alerts
- pkg/notify/jira.go: Infof → Infow (log injection)
- service/sync_status.go: uint→int → uint→int64 (integer overflow)
- pkg/otel/attrs.go: add nolint comment for safe uint→int conversion

Remaining alerts are false positives:
- go/request-forgery in cf/client.go: endpoint validated to *.hana.ondemand.com
- go/log-injection in service/: all use structured w-variants with uint/enum types

+++ service/sync_status.go
@@ -95,7 +95,7 @@ func (s *Service) DetermineOverallStatus(drID uint) error {
 
 func (s *Service) SyncDeliveryStatus(ctx context.Context, deliveryRequestID uint, user string) error {
 	ctx, span := cpiotel.Tracer().Start(ctx, "SyncDeliveryStatus",
-		oteltrace.WithAttributes(attribute.Int("dr_id", int(deliveryRequestID))))
+		oteltrace.WithAttributes(attribute.Int64("dr_id", int64(deliveryRequestID))))
 	defer span.End()
 
 	// Guard against concurrent sync for the same DR.
